requirements-docker.txt
# FY3 Automation Center — Docker (web-app only, no torch/librosa)
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
aiohttp>=3.10.0
aiofiles>=24.1.0
httpx>=0.27.0
pydantic>=2.10.0
pyyaml>=6.0
python-dotenv>=1.0.0
python-multipart>=0.0.18
Pillow>=11.0.0
google-api-python-client>=2.160.0
google-auth>=2.36.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
requests>=2.32.0
websockets>=14.0
